Completed the Blog Form page with proper popup and validation

+++ src/main/java/com/example/travellerblog/service/ImageStorageServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.multipart.commons.CommonsMultipartFile;
 
+import javax.annotation.PostConstruct;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -20,13 +21,15 @@ public class ImageStorageServiceImpl implements ImageStorageService{
     private final Path imageDir = Paths.get("images");
 
     @Override
+    @PostConstruct
     public void init() throws IOException {
-        Files.createDirectory(imageDir);
+        if (!Files.exists(imageDir)) Files.createDirectory(imageDir);
     }
 
     @Override
     public String saveImage(MultipartFile file) throws IOException {
-        Files.copy(file.getInputStream(), this.imageDir.resolve(Objects.requireNonNull(file.getOriginalFilename())));
+        if (!Files.exists(this.imageDir.resolve(Objects.requireNonNull(file.getOriginalFilename()))))
+            Files.copy(file.getInputStream(), this.imageDir.resolve(Objects.requireNonNull(file.getOriginalFilename())));
         return this.imageDir.resolve(file.getOriginalFilename()).toString();
     }
